refactor(bridge): extract shared HTTP retry/backoff/timeout helpers
Both bridge artifacts duplicated the retry/backoff/timeout primitives:
constants (MAX_RETRIES, INITIAL_BACKOFF_MS, FETCH_TIMEOUT_MS,
DELETE_MAX_RETRIES, DELETE_BACKOFF_MS), the AbortController-based
fetchWithTimeout, and the exponential backoff math + sleep.

Extract those identical primitives into bridge/src/http-retry.ts.
- bridge imports relatively (./http-retry.js)
- bridge-app imports via the existing @bridge/* alias
  (@bridge/http-retry.js), mirroring @bridge/circuit-breaker.js

The retry LOOPS stay in each artifact: they diverge legitimately
(bridge-app records backendReachable/status transitions and stops the
bridge on a 401 token expiry; the headless bridge does neither). That
divergence is intentionally NOT folded into the shared helper.

Behavior-preserving: constants identical, circuit-breaker config
unchanged, 401 stop path intact. tsc + tests green in both packages;
Electron Vite production bundle builds and externalizes correctly.

Closes #531
